fixed issue with payment info section and form validation
Fixed these issues:

Payment Info Section
There seems to be a problem with the "Payment" info sections. If PayPal or Bitcoin are selected the form will submit if just a single activity is selected and nothing else on the page is filled in.

Form Validation
Currently if everything else on the page is filled out, only a single credit card input needs to be filled in and the form will submit.